KU_Certification/requirements.txt
Flask==2.0.1
Flask-Session==0.4.0
gunicorn==20.1.0
Werkzeug==2.0.1
psycopg2-binary==2.9.1
Flask-WTF==1.0.1